Fix summary results
When submitting, it will loop through the number of days and determine the summary. There was an issue with the values not being parsed properly so that was fixed by adding a step to the inputs, and then parsing the inputs only when needed. Tax percentage still needs to be included.